ml_recsys/requirements.txt
torch
numpy
pandas
scikit-learn
faiss-cpu